Introduce RISC-V support in goodasm
The language support files are duplicated -
galangriscv.{cpp|h} and galangriscv32.{cpp|h}.
This is with the intention of follow-up work
to add galangriscv64.{cpp|h}, which the RISCV
spec identifies as a distinct base ISA.

To suppport or streamline compressed instruction
extension encoding, a new feature was added
to goodasm.cpp for rejectWhenZero constraint
matching. This makes it easier to differentiate
between some compressed instructions against c.nop
or each other. There may be ways to resolve and
improve the bitmask / match() without this,
but did not appear to impact performance.

(An alternative path is to generalize
rejectWhenZero so that you can support
arbitrary constraint matching - this
might help with other languages with
funky edge-cases in the future.)

A large number of test cases were added to
riscv/tests - including an instruction
generation python script that attempted to
bruteforce every possible valid instruction.
The general testing approach was reflexive -
encode .asm to binary, decode that binary
to .disasm. If .asm and .disasm match,
it increases confidence of the correct
encoding. Formatting need{s/ed} to be accounted
for and the possibility of incorrect
binary representations remain - more robust
testing would include QEMU virtualization.

Various AI coding assistants were used in
combination with manual effort in an attempt
to explore LLM capabilities; this contribution
is "spec-driven development" by definition.
There is a chance errors slipped through despite
best efforts to manually audit and refine.